==> ecommerce_customer_churn_prevention/notebooks/1.0-wfp-etl-process.ipynb <==
# Preparing the environment

import pandas as pd
import numpy as np
import sys
sys.path.append('../ecommerce_customer_churn_prevention')
from utils import check_typo_errors, paths
import warnings
warnings.filterwarnings(category=FutureWarning, action='ignore')

# Data dictionary

data_dict = pd.read_excel(paths.data_raw_dir('e_commerce_dataset.xlsx'), sheet_name=0, header=1)
data_dict

# Importing the data


df = pd.read_excel(paths.data_raw_dir('e_commerce_dataset.xlsx'), sheet_name=1)
df.head()

# Converting the features to categorical like the data dictionary

cat_features = ['Churn', 'PreferredLoginDevice', 'CityTier', 'PreferredPaymentMode', 'Gender', 'PreferedOrderCat', 'MaritalStatus', 'Complain']

df[cat_features] = df[cat_features].astype('category')

df.info()

# Looking for missing values


df.isna().sum()

Missing values are only present in numerical features. Which makes easy to impute them using descriptive statistics or imputation methods like KNNImputer.

# Looking for typo errors in the categorical features


for col in cat_features:
    check_typo_errors.check_categories_proportion(df, col)

PreferredLoginDevice has a typo error in the category 'Phone'. 'Phone' should be 'Mobile Phone' as well. Also, in the feature PreferredPaymentMode, the categories COD and CC are Cash on Delivery and Credit Card respectively, so they will be renamed to Cash on Delivery and Credit Card. Finally, "Mobile" in the feature 'PreferedOrderCat' should be 'Mobile Phone' as well.


df['PreferredLoginDevice'] = df['PreferredLoginDevice'].replace({'Phone': 'Mobile Phone'})
df['PreferredPaymentMode'] = df['PreferredPaymentMode'].replace({'COD': 'Cash on Delivery', 'CC': 'Credit Card'})
df['PreferedOrderCat'] = df['PreferedOrderCat'].replace({'Mobile':'Mobile Phone'})

for col in cat_features:
    check_typo_errors.check_categories_proportion(df, col)

# Checking descriptive statistics


df.describe().T

There are outliers in the features 'Tenure', 'WarehouseToHome', 'NumberOfAddress', 'CouponUsed', 'OrderCount' and 'DaySinceLastOrder'. There are possible outliers in the features 'OrderAmountHikeFromLastYear'.

They will be explored and treated in the Exploratory Data Analysis phase.


df.describe(include='category').T

The feature 'Churn' has a high imbalance between the categories. The dataset is unbalanced. So, the model must be robust to the imbalance. A good option is to use XGBoostClassifier with the parameter scale_pos_weight, focusing on the F1 score since the company is interested in minimizing the false negatives while maximizing the true positives.


# Exploring null values


df.isna().sum().to_frame(name='Null values')


df.isna().sum(axis=1).value_counts().to_frame(name='Null values per row')


df.isna().sum(axis=1).value_counts(normalize=True).mul(100).to_frame(name='Null values per row proportion')


All the null values are in different rows and different columns. They also are present in the 33% of the rows, so dropping them is not a good option. Since XGBoostClassifier can handle null values, they will not be imputed.


df.to_csv(paths.data_interim_dir('df_etl_processed.csv'), index=False)
