==> ecommerce_customer_churn_prevention/notebooks/3.0-wfp-hypothesis-testing.ipynb <==
# Hypothesis testing

In statistics, is not enough with exploring the data visually, but it's necessary to make statistical test over them to confirm or negate our hypotheses. In this case, we gonna focus on non-parametric statistical tests, such as Kruskal-Wallis and Chi-Square tests, to test the relationship between the numerical and categorical variables with our objective variable. 

# Preparing the environment

import pandas as pd
import sys
sys.path.append('../ecommerce_customer_churn_prevention')
from utils import paths, hypothesis_testing

# Importing the data

df = pd.read_csv(paths.data_interim_dir('df_etl_processed.csv'))
df.head()

# Converting the features to categorical like the data dictionary

cat_features = ['PreferredLoginDevice', 'CityTier', 'PreferredPaymentMode', 'Gender', 'PreferedOrderCat', 'MaritalStatus', 'Complain']
num_features = [col for col in df.columns if col not in cat_features]
num_features.remove('Churn')
num_features.remove('CustomerID')

df[cat_features] = df[cat_features].astype('category')

df.info()

# Applying hypothesis testing

sig_cols = []
no_sig_cols = []
for col in cat_features:
    hypothesis_testing.chi2_independence_test(df, 'Churn', col, sig_col=sig_cols, no_sig_col=no_sig_cols)

for col in num_features:
    hypothesis_testing.kruskal_wallis_test(df, 'Churn', col, sig_col=sig_cols, no_sig_col=no_sig_cols)

print(f"Significant columns: {sig_cols}")
print(f"No significant columns: {no_sig_cols}")
