==> ecommerce_customer_churn_prevention/notebooks/4.0-wfp-feature-engineering.ipynb <==
# Feature Engineering

Feature engineering is the art of creating new features based on the available ones in order to increase model performance. In this case, feature engineering will be realized using the hypothesis testing results, looking for catch complex relationships between the features that most influence in the churning of the customers.

# Preparing the environment

import pandas as pd
import sys
sys.path.append('../ecommerce_customer_churn_prevention')
from utils import paths

# Importing the data

df = pd.read_csv(paths.data_interim_dir('df_etl_processed.csv'))
df.head()

# Converting the features to categorical like the data dictionary

cat_features = ['PreferredLoginDevice', 'CityTier', 'PreferredPaymentMode', 'Gender', 'PreferedOrderCat', 'MaritalStatus', 'Complain']

df[cat_features] = df[cat_features].astype('category')

df.info()

# Interactions between categorical features

df['PreferredLoginDevice_CityTier'] = df['PreferredLoginDevice'].astype(str) + '_' + df['CityTier'].astype(str)
df['PreferredLoginDevice_PreferredPaymentMode'] = df['PreferredLoginDevice'].astype(str) + '_' + df['PreferredPaymentMode'].astype(str)
df['CityTier_PreferredPaymentMode'] = df['CityTier'].astype(str) + '_' + df['PreferredPaymentMode'].astype(str)
df['Gender_PreferedOrderCat'] = df['Gender'].astype(str) + '_' + df['PreferedOrderCat'].astype(str)
df['Gender_MaritalStatus'] = df['Gender'].astype(str) + '_' + df['MaritalStatus'].astype(str)
df['Gender_Complain'] = df['Gender'].astype(str) + '_' + df['Complain'].astype(str)
df['PreferedOrderCat_MaritalStatus'] = df['PreferedOrderCat'].astype(str) + '_' + df['MaritalStatus'].astype(str)
df['Complain_MaritalStatus'] = df['Complain'].astype(str) + '_' + df['MaritalStatus'].astype(str)
df['Complain_PreferedOrderCat'] = df['Complain'].astype(str) + '_' + df['PreferedOrderCat'].astype(str)

# Interactions between numerical features

df['OrderCount_Tenure_Ratio'] = df['OrderCount'] / (df['Tenure'] + 1)
df['CouponUsed_Tenure_Ratio'] = df['CouponUsed'] / (df['Tenure'] + 1)
df['WarehouseToHome_Tenure_Ratio'] = df['WarehouseToHome'] / (df['Tenure'] + 1)
df['HourSpendOnApp_Tenure_Ratio'] = df['HourSpendOnApp'] / (df['Tenure'] + 1)
df['SatisfactionScore_Tenure_Ratio'] = df['SatisfactionScore'] / (df['Tenure'] + 1)
df['HourSpendOnApp_NumberOfDeviceRegistered'] = df['HourSpendOnApp'] * df['NumberOfDeviceRegistered']
df['SatisfactionScore_NumberOfDeviceRegistered'] = df['SatisfactionScore'] * df['NumberOfDeviceRegistered']
df['SatisfactionScore_OrderCount'] = df['SatisfactionScore'] * df['OrderCount']
df['SatisfactionScore_CouponUsed'] = df['SatisfactionScore'] * df['CouponUsed']
df['SatisfactionScore_CashbackAmount'] = df['SatisfactionScore'] * df['CashbackAmount']
df['CashbackAmount_Tenure_Ratio'] = df['CashbackAmount'] / (df['Tenure'] + 1)
df['CashbackAmount_CouponUsed_Ratio'] = df['CashbackAmount'] / (df['CouponUsed'] + 1)

# Saving the processed data with new features

df.to_csv(paths.data_processed_dir('df_processed.csv'), index=False)
