==> ecommerce_customer_churn_prevention/notebooks/6.1-wfp-model-optimization-2.ipynb <==
# Model Optimization

In this case, a model of the data with the total of the selected features in the dimensionality reduction will be optimized, but with the first 18 features that led the model to have F1 score above 0.9, having the goal to check if these reduced features will improve or at least have the same performance of the previous model and reduce overfitting. Optimization of the model will be focused on F1 score since our data is imbalanced, and we want to have a good prediction from both minority and majority class.

The hyperparameter tuning will be made using bayesian optimization, which is a kind of "informed optimization" using a probabilistic approach in order to reach the minimum of the loss function score. The framework "optuna" is ideal for this bayesian optimization.

# Preparing environment

import pandas as pd
import xgboost as xgb
import optuna
from sklearn.model_selection import train_test_split
from sklearn.metrics import f1_score, confusion_matrix
import sys
sys.path.append('../ecommerce_customer_churn_prevention')
from utils import paths
import warnings
warnings.filterwarnings('ignore')

# Importing the data

X = pd.read_csv(paths.data_processed_dir('df_red_selected_features.csv'))
y = pd.read_csv(paths.data_processed_dir('df_processed.csv'))['Churn']
X.head()

# Converting the features to categorical like the data dictionary

cat_features = ['PreferedOrderCat', 'Complain']

X[cat_features] = X[cat_features].astype('category')

# Converting the new features to categorical
new_feat_cat = [col for col in X.select_dtypes('object').columns]
X[new_feat_cat] = X[new_feat_cat].astype('category')

X.info()

# Dividing data into train and test

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42, stratify=y)

# Defining function to optimize

def objective(trial):
    # Suggest hyperparameters
    params = {
        'objective': 'binary:logistic',
        'eval_metric': 'logloss',
        'use_label_encoder': False,
        'learning_rate': trial.suggest_float('learning_rate', 0.01, 0.3),
        'max_depth': trial.suggest_int('max_depth', 3, 15),
        'min_child_weight': trial.suggest_int('min_child_weight', 1, 10),
        'subsample': trial.suggest_float('subsample', 0.5, 1.0),
        'colsample_bytree': trial.suggest_float('colsample_bytree', 0.5, 1.0),
        'gamma': trial.suggest_float('gamma', 0, 5),
        'scale_pos_weight': trial.suggest_float('scale_pos_weight', 1.0, 10.0),
        'lambda': trial.suggest_float('lambda', 0.0, 10.0),
        'alpha': trial.suggest_float('alpha', 0.0, 10.0),
    }
    
    # Train/Validation split
    X_train_f, X_val, y_train_f, y_val = train_test_split(X, y, test_size=0.25, stratify=y)
    
    # Train XGBoost
    dtrain = xgb.DMatrix(X_train_f, label=y_train_f, enable_categorical=True)
    dval = xgb.DMatrix(X_val, label=y_val, enable_categorical=True)
    model = xgb.train(params, dtrain, evals=[(dval, "validation")], verbose_eval=False, num_boost_round=100, early_stopping_rounds=10)
    
    # Predictions
    preds = model.predict(dval)
    preds = (preds > 0.5).astype(int)
    
    # Evaluate
    return f1_score(y_val, preds)

# Create and optimize the study

study = optuna.create_study(direction="maximize")
study.optimize(objective, n_trials=100)

# Best parameters
print("Best Parameters:", study.best_params)

# Training the final model and testing on test data

# Training will be made with both train and val data, then tested on test data

best_params = study.best_params
xgb_model = xgb.XGBClassifier(objective='binary:logistic', eval_metric='logloss', use_label_encoder=False, enable_categorical=True, **best_params)
xgb_model.fit(X_train, y_train, verbose=0)
y_pred = xgb_model.predict(X_test)

# Testing the model

print(f'F1 Score for model optimized with Optuna and all the selected categories: {f1_score(y_test, y_pred)}')
print(f'Confusion matrix: \n{confusion_matrix(y_test, y_pred)}')

The model has lower F1 Score than the previous model. Nontheless, 0.8985 F1 score with 4 false positives and 38 false negatives indicates a better performance on this model for predicting the customers who will really churn, since the complexity of the data was reduced and the patterns of the churn customers were clearer.

The decision of work with any of the models is on hands of the stakeholders, since they decide how to retain the customers and their strategy from this analysis, which may involves discounts and some other promotions for the churning customers predicted.

# Saving the model
xgb_model.save_model(paths.models_dir('xgb_model_optuna_red.json'))
